etl_web_platform/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
waitress==2.1.2
psutil==5.9.5
pandas==2.0.3
sqlalchemy==2.0.20
pydantic==2.3.0
numpy==1.24.3
dask==2023.8.0
bcrypt==4.0.1
